Move actor editing form under editing item
